backend/requirements.txt
Django==4.1.5
django-redis==5.2.0
django-simpleui==2022.12.18
PyJWT==2.6.0
PyMySQL==1.0.2
rich==13.0.1
six==1.16.0
geoip2==4.6.0
psutil==5.9.4
timeloop==1.0.2
redis==4.4.1
django-hcaptcha==0.2.0
oauth2==1.9.0.post1
numpy==1.24.1
pillow==9.4.0
django-imagekit==4.1.0
django-haystack==3.2.1
jieba==0.42.1
whoosh==2.7.4
gevent==22.10.2
gunicorn==20.1.0
websockets===10.4
threadpool===1.3.2
markdown==3.4.1
django-mdeditor==0.1.20
django-mptt==0.14.0
paddlepaddle==2.4.1
paddlehub==2.3.1
eventlet==0.33.2
celery==5.2.7
